pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vtts"
version = "0.1.0"
description = "Virtual TTS: continuous batching for text-to-speech, like vLLM for voice"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Antoni Caimari Caldes", email = "acaimari22@gmail.com"},
]
keywords = ["tts", "text-to-speech", "continuous-batching", "qwen3", "voice-cloning", "streaming"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=2.1.0",
    "transformers>=4.40.0",
    "numpy",
    "soundfile",
    "pyyaml",
    "fastapi",
    "uvicorn",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "httpx",
]

[project.urls]
Homepage = "https://github.com/caimari/vtts"
Repository = "https://github.com/caimari/vtts"

[tool.setuptools.packages.find]
include = ["vtts*"]
